feat: add remote control support for 'common' ext
My idea here is to at one point move both the manager and the hardcoded
IP address to sophys-common or similar, such that this can go into
pre-execution and be generic to all beamlines, similar to what we do with
the kafka callbacks.

But for now, a bit of hardcoding and duplication is needed.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>