fix: prevent kbl from quitting when cancelling an authorization request
Previously the kbl process would be in the same process group as the
parent process, thus the SIGINT signal would be forwarded by the kernel
to all of them. In the parent process, we handle that in a try..except
block, but in the kbl process no.

Thus, we separate them into different sessions, so signals are not forwarded.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>