refactor: move PlanWhitelist objects to their own special class
This is made to work around an issue with the old approach, where
we couldn't have two user different plans defined from a single
upstream plan.

This will be required in EMA's case, where we'll have absolute
scans and relative scans only change by one parameter, on the client.

This seems easier to manage overall, though it would be possible to
create separate plans on sophys-ema instead.

This also makes the PlanWhitelists more extensible and better typed,
which is also why I decided on this approach.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>