feat: add test CLI flag
This is intended to provide a way for us to more easily create a testing
environment for feature validation.

It does NOT deal with testing at the same time people use the tool in production,
as they will still be using the same RE. What is does is only provide a separation
of concerns at the level of global variables and parameters, like Redis key values.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>