refactor: make 'pre_execution' an importable module
Now, 'base_configuration.py' is a normal Python module, which
can be imported at will, and won't run anything from that. Only
'_pre_execution.py' will run something, so it gets easier to
add tests to all that functionality later on.

The 'globals' must be passed on because the reference global
scope used throughout the application is the one from
'_pre_execution', so if we don't do that, the things we add to
the namespace will not be accessible elsewhere.

I've also switched the 'common' default extension name to 'skip',
so we don't try to connect to Kafka in the tests anymore.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>