refactor+fix: switch _pre_execution to an extension loading
The technical motivation for this change is that, before, all the
extensions would load **before** execute_at_start ran, because that's
the initialization ordering in TerminalIPythonApp. This meant that
extensions would not be able to access objects created during that
setup.

This patch makes the basic setup an extension as well, and we rely on
the fact that extension loading happens in-order to ensure the ordering
is as we would expect it.

The less technical reasoning for this change is that by doing it like so,
we use only a single concept (extensions) to load all the code required
by sophys-cli, instead of relying on two mechanisms.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>