test: use exec instead of locals() manipulation in magic calls
In Python 3.13, the semantics of locals() became well-defined, and
in optimized scopes (like functions), we do not get a reference to
the frame locals anymore, we get a copy. This means we cannot rely
on mutating it to setup the local environment.

So, we can instead setup the namespace ourselves and call exec on
that new namespace, since we're only interested in the side effects.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>